Show indication of wrong or right answer in question feedback
- v and x -signs have been added to hints displaying at the end
of each question.
- If there are no hints to be shown with either the correct or false
answer, a default-message is shown to more clearly indicate to
the student which answer is correct and which not. This is also to
improve the accessibility of the correctness.
- Checkboxes are handled a bit differently, the indication of correctness
is not done hint-wise as it would benefit the student too much but instead
at the end of the question
- If there a choices in a question, the chosen one has been highlighted
- All the previous indications of correctness have been preserved.