Show indication of wrong or right answer in question feedback
- v and x -signs have been added to hints displaying at the end
of each question.
- If there are no hints to be shown with either the correct or false
answer, a default-message is shown. This is done to indicate to
the student which answer is correct and which not, and to not make
the questionnaire rely so heavily on just for color.
- Checkboxes are handled a bit differently, the indication of correctness
is NOT done hint-wise. That would benefit the student too much.
- If there a choices in a question, the chosen one has been highlighted
- All the previous indications of correctness have been preserved so that
different yaml-configurations still work.