Clean up the feedback when model answers are shown
- After the student has reached max-submissions, if the answer is not
correct, but staff has set the model to be visible for the student,
correct answers are shown.
- The html-elements with "data-aplus-once" are only to be shown
when the data is first stored to the database. Thus correct answers
cannot be checked later on, only right after the max submissions have
been reached. This has not been changed from the previous implementation.